aws_lambda_functions/index_papers/requirements.txt
farm-haystack[preprocessing,faiss,inference,file-conversion,pdf]
awslambdaric
PyPDF2
boto3